Refactor to simplify package structure
