+++ src/main/java/org/jboss/maven/extension/dependency/metainf/MetaInfGenerator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.maven.extension.dependency.metainf.generator;
+package org.jboss.maven.extension.dependency.metainf;
 
 import java.io.IOException;
