+++ src/main/java/org/jboss/maven/extension/dependency/modelmodifier/versionoverride/DepVersionOverrider.java
@@ -28,9 +28,9 @@ import org.apache.maven.model.DependencyManagement;
 import org.apache.maven.model.Model;
 import org.apache.maven.model.building.ModelBuildingException;
 import org.jboss.maven.extension.dependency.resolver.EffectiveModelBuilder;
+import org.jboss.maven.extension.dependency.util.Log;
 import org.jboss.maven.extension.dependency.util.MavenUtil;
 import org.jboss.maven.extension.dependency.util.VersionPropertyReader;
-import org.jboss.maven.extension.dependency.util.log.Log;
 import org.sonatype.aether.resolution.ArtifactDescriptorException;
 import org.sonatype.aether.resolution.ArtifactResolutionException;
