+++ src/main/java/org/jboss/maven/extension/dependency/util/Log.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.maven.extension.dependency.util.log;
+package org.jboss.maven.extension.dependency.util;
 
 import org.codehaus.plexus.logging.Logger;
 import org.codehaus.plexus.logging.console.ConsoleLogger;
